+++ RbacWebModule.php
@@ -17,13 +17,8 @@ use yii\filters\AccessControl;
 /**
  * @author Dmitry Erofeev <dmeroff@gmail.com>
  */
-class Module extends BaseModule
+class RbacWebModule extends BaseModule
 {
-    /**
-     * @var bool Whether to show flash messages
-     */
-    public $enableFlashMessages = true;
-
     /**
      * @var string
      */