+++ views/rule/index.php
@@ -10,18 +10,19 @@
  */
 
 /**
- * @var $dataProvider array
- * @var $this         yii\web\View
- * @var $filterModel  dektrium\rbac\models\Search
+ * @var $this         \yii\web\View
+ * @var $searchModel  \dektrium\rbac\models\RuleSearch
+ * @var $dataProvider \yii\data\ArrayDataProvider
  */
 
 use kartik\select2\Select2;
 use yii\grid\ActionColumn;
 use yii\grid\GridView;
 use yii\helpers\Url;
+use yii\web\JsExpression;
 use yii\widgets\Pjax;
 
-$this->title = Yii::t('rbac', 'Permissions');
+$this->title = Yii::t('rbac', 'Rules');
 $this->params['breadcrumbs'][] = $this->title;
 
 ?>
@@ -32,62 +33,65 @@ $this->params['breadcrumbs'][] = $this->title;
 
 <?= GridView::widget([
     'dataProvider' => $dataProvider,
-    'filterModel'  => $filterModel,
+    'filterModel'  => $searchModel,
     'layout'       => "{items}\n{pager}",
     'columns'      => [
         [
             'attribute' => 'name',
-            'header'    => Yii::t('rbac', 'Name'),
+            'label'     => Yii::t('rbac', 'Name'),
             'options'   => [
                 'style' => 'width: 20%'
             ],
             'filter' => Select2::widget([
-                'model'     => $filterModel,
+                'model'     => $searchModel,
                 'attribute' => 'name',
-                'data'      => $filterModel->getNameList(),
-                'options'   => [
-                    'placeholder' => Yii::t('rbac', 'Select permission'),
-                ],
-                'pluginOptions' => [
-                    'allowClear' => true,
-                ],
-            ]),
-        ],
-        [
-            'attribute' => 'description',
-            'header'    => Yii::t('rbac', 'Description'),
-            'options'   => [
-                'style' => 'width: 55%',
-            ],
-            'filterInputOptions' => [
-                'class'       => 'form-control',
-                'id'          => null,
-                'placeholder' => Yii::t('rbac', 'Enter the description')
-            ],
-        ],
-        [
-            'attribute' => 'rule_name',
-            'header'    => Yii::t('rbac', 'Rule name'),
-            'options'   => [
-                'style' => 'width: 20%'
-            ],
-            'filter' => Select2::widget([
-                'model'     => $filterModel,
-                'attribute' => 'rule_name',
-                'data'      => $filterModel->getRuleList(),
                 'options'   => [
                     'placeholder' => Yii::t('rbac', 'Select rule'),
                 ],
                 'pluginOptions' => [
+                    'ajax' => [
+                        'url'      => Url::to(['search']),
+                        'dataType' => 'json',
+                        'data'     => new JsExpression('function(params) { return {q:params.term}; }')
+                    ],
                     'allowClear' => true,
+                    
                 ],
             ]),
+        ],
+        [
+            'attribute' => 'class',
+            'label'     => Yii::t('rbac', 'Class'),
+            'value'     => function ($row) {
+                $rule = unserialize($row['data']);
+
+                return get_class($rule);
+            },
+            'options'   => [
+                'style' => 'width: 20%'
+            ],
+        ],
+        [
+            'attribute' => 'created_at',
+            'label'     => Yii::t('rbac', 'Created at'),
+            'format'    => 'datetime',
+            'options'   => [
+                'style' => 'width: 20%'
+            ],
+        ],
+        [
+            'attribute' => 'updated_at',
+            'label'     => Yii::t('rbac', 'Updated at'),
+            'format'    => 'datetime',
+            'options'   => [
+                'style' => 'width: 20%'
+            ],
         ],
         [
             'class'      => ActionColumn::className(),
             'template'   => '{update} {delete}',
             'urlCreator' => function ($action, $model) {
-                return Url::to(['/rbac/permission/' . $action, 'name' => $model['name']]);
+                return Url::to(['/rbac/rule/' . $action, 'name' => $model['name']]);
             },
             'options'   => [
                 'style' => 'width: 5%'